+++ client/fluence-ipfs/Cargo.toml
@@ -9,8 +9,8 @@ fluence-libp2p = { path = "../../crates/libp2p" }
 fluence-client = { path = "../rust-libp2p" }
 faas-api = { path = "../../faas-api" }
 ctrlc-adapter = { path = "../../crates/ctrlc-adapter"}
-libp2p = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "deserialize" }
-parity-multiaddr = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "deserialize" }
+libp2p = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "master" }
+parity-multiaddr = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "master" }
 multihash ="0.11.0"
 async-std = "1.5.0"
 futures = "0.3.4"
@@ -28,6 +28,5 @@ itertools = "0.9.0"
 uuid = { version = "0.8.1", features = ["serde", "v4"] }
 async-timer = { version = "0.7.3", features = ["stream"] }
 once_cell = "1.3.1"
-clap = "2.33"
 
 [dev-dependencies]
