+++ crates/libp2p/Cargo.toml
@@ -5,8 +5,8 @@ authors = ["Fluence Labs"]
 edition = "2018"
 
 [dependencies]
-libp2p = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "deserialize" }
-parity-multiaddr = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "deserialize" }
+libp2p = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "master" }
+parity-multiaddr = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "master" }
 multihash ="0.11.0"
 futures = "0.3.4"
 futures-util = "0.3.4"
