+++ deploy/fluence_bootstrap.yml
@@ -1,7 +1,7 @@
 version: '3.5'
 services:
   fluence_bootstrap:
-    image: fluencelabs/fluence:routing
+    image: fluencelabs/fluence:master
     container_name: bootstrap
     restart: always
     command: "-c /Config.toml -s /node.key -e $HOST -t 7770 -w 9990"
@@ -13,7 +13,7 @@ services:
       RUST_LOG: "debug,tokio_threadpool=info,tokio_reactor=info,mio=info,tokio_io=info,soketto=info,yamux=info,multistream_select=info,libp2p_secio=info,libp2p_websocket::framed=info,libp2p_ping=info,libp2p_core::upgrade::apply=info,libp2p_kad::kbucket=info"
 
   fluence-ipfs:
-    image: fluencelabs/fluence:routing-with-ipfs-multiaddr
+    image: fluencelabs/fluence:master-with-ipfs-multiaddr
     restart: always
     ports:
       - "7100:7100" # Node service
