+++ faas-api/Cargo.toml
@@ -7,8 +7,8 @@ edition = "2018"
 [dependencies]
 trust-graph = { path = "../trust-graph" }
 fluence-libp2p = { path = "../crates/libp2p" }
-libp2p = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "deserialize" }
-parity-multiaddr = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "deserialize" }
+libp2p = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "master" }
+parity-multiaddr = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "master" }
 multihash ="0.11.0"
 futures = "0.3.4"
 futures-util = "0.3.4"
