+++ server/Cargo.toml
@@ -9,7 +9,7 @@ trust-graph = { path = "../trust-graph" }
 fluence-libp2p = { path = "../crates/libp2p" }
 ctrlc-adapter = { path = "../crates/ctrlc-adapter"}
 faas-api = { path = "../faas-api" }
-fluence-ipfs = { path = "../client/fluence-ipfs" }
+fluence-faas = { git = "https://github.com/fluencelabs/fce", branch = "integration" }
 libp2p = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "master" }
 parity-multiaddr = { git = "https://github.com/fluencelabs/rust-libp2p", branch = "master" }
 multihash ="0.11.0"
@@ -19,7 +19,6 @@ async-std = "1.5.0"
 itertools = "0.9.0"
 serde = { version = "1.0.110", features = ["derive"] }
 serde_json = "1.0.53"
-serde_with = { version = "1.4.0", features = ["json"] }
 bs58 = "0.3.0"
 failure = "0.1.6"
 base64 = "0.11.0"
