+++ server/src/lib.rs
@@ -26,12 +26,14 @@
     unreachable_patterns
 )]
 
+/// Takes option as an argument, unwraps if `Some`, exit function with `Ok(default)` otherwise
+/// Ought to make it easier to short-circuit in functions returning `Result<Option<_>>`
 macro_rules! ok_get {
     ($opt:expr) => {{
         let r = { $opt };
         match r {
-            None => return Ok(Default::default()),
             Some(r) => r,
+            None => return Ok(Default::default()),
         }
     }};
 }
